+++ src/main/java/br/com/modtiktok/tiktokchaos/preset/PresetCompatibility.java
@@ -1,0 +1,6 @@
+package br.com.modtiktok.tiktokchaos.preset;
+
+import java.util.List;
+
+public record PresetCompatibility(boolean available, List<String> missingRequirements) {
+}